up date example and add some written docs
