+++ core/dbt/cli/__init__.py
@@ -1,0 +1,1 @@
+from .main import cli as dbt_cli  # noqa
